Add absolute body fat and muscle on DataEntry, Overview and Table
... so that the users can also monitor whether they lose or gain fat
or muscle when their weight changes.

Also change styling of fat and muscle icons: they now have a percentage
sign whereas the absolute icons show the original ones.
I removed the sesame from the burger so that the percentage sign is
better recognized. =)

Please check the french translation and provide the japanese one.