Add support for non-binary gender
The ScaleUser.gender variable can now be set to 2 to indicate a non-binary gender, and a corresponding radio button has been added to the UI.  For evaluation sheets, we evaluate a non-binary user as "NORMAL" if their body metrics are within the combined NORMAL ranges for male and female metrics.  For body metrics that depend on biological sex, we return the average of the male and female values for a non-binary user.  For communicating with bluetooth scales, we largely stick to the existing code pattern, designating all non-male users as female.

This is kind of terrible, but I don't have any better ideas for how to make this app as gender-inclusive as possible while relying on data from non-inclusive research papers and interfacing with non-inclusive external devices.