Avoid updating diff value twice
As of 60471ab41, the diff value is updated when the value changes. But
if we know that the diff value will be updated after the value anyway
we should not update the diff value.

Introduce a reset method that clears the value and diff value so that
when the value is set, the diff value will not be updated.