Show days since current date in overview
I find it more logical to see how long ago a measurement was done,
instead of how long ago since the last measurement was a measurement
done.

Fix a bug where the days since a measurement was calculated
incorrectly if the measurement was done in the previous year.

Convert the days label to plurals so that "day" is used when the
measurement was done yesterday.