Postpone creating text until view is requested
Wait until the measurement is actually to be shown in the gui until
creating the spanned text. Cache it so that it doesn't have to be
recalculated.

Drop the first hidden column that contained the id. Instead implement
Adapter.getItemId and use the supplied id in onClickListenerRow.

Speeds up the initial loading of the table.