Don't show soft input by default in input dialog
It takes time to open the soft input and with the new inc/dec buttons
it should not be needed as much. If needed it is quickly summoned by
clicking the input view.

Also it doesn't look nice when the phone is in horizontal mode as then
the soft input takes up the whole screen and one has to push back to
get to the inc/dec buttons.

See discussion in #222.