Add language preference
Make it possible to select a language that differs from the system
default. This can be useful when doing translations (as hinted in
issue #186), or if you just want to use openScale in a different
language from the system one.

Also moved theme setting to a new common base class for activities.