Do BLE scan in addition to bluetooth discovery
After running the old school bluetooth discovery, start a 10 second
BLE scan. This is a better way to find BLE devices and at least for me
always find my scale (something that the discovery doesn't always do).

Also add a simple progress bar that is shown while the discovery is in
progress.

Finally add a link to the wiki for users with unsupported scales.