Add bluetooth debug driver
The driver reads all readable characteristics and all descriptors from
the device when connected and logs them. The mode is enabled when
saving log to file is enabled and can be used on unsupported devices
as well.

This replaces the need to use the BLE scanner with something that
should be more easy to use.