Remove user initial weight
As stated in #253 it is confusing for a first time user to have to
enter an initial weight when adding an user.

The comment from a first time user was something like "Why do I need
to enter this? I don't know what my weight is. That's what the scale
is supposed to tell me!".