Add measurement widget
The widget can be added to the home screen and can be configured to
show any of the measurements that are enabled. It can also be resized
from 1x1 to <max width>x1, adapting to the size change and showing a
reasonable amount of information at each size. The default size is
4x1.