Run a LE scan for the device before connecting
Seems to be a requirement for some scales (e.g. Medisana BS444) and we
might as well do it for all scales since it should hopefully improve
connectivity for others as well.

If no location permission is given, or if no device is found within 10
seconds, we try to connect directly to the scale.

As of this, #278 should be fixed.