Add a ContentProvider so other apps can access (with permission) the openScale data.
This could be used for e.g. syncing weight logs to a third-party service such as Google Fit,
without the openScale app having to do so; with this, a 'connector' app could be written that
reads this data and does the push to a third-party service (or, indeed, an HTTP RPC connector,
an MQTT connector, SMS notifications, or anything else) without openScale needing this
functionality (or additional permissions).