BeurerSanitas: Retry step 4 when interrupted by unexpected messages
The scale has a habit of sending the status message multiple times, even after being asked for the user list.
When this happens the state machine must be reset to state 4 and retry.

Similarly, the scale sends our measurements when the user is standing on it during the init process.
If this happens in state 4, we have to try again.